ON-9609 # Refactored OptionButton to component
